primitives, state: moved EpochCommitment to primitives crate
