+++ crates/primitives/src/lib.rs
@@ -7,14 +7,13 @@ pub mod block_credential;
 pub mod bridge;
 pub mod buf;
 pub mod constants;
+pub mod epoch;
 pub mod errors;
 pub mod evm_exec;
 pub mod hash;
+pub mod keys;
 pub mod l1;
 pub mod l2;
-#[macro_use]
-mod macros;
-pub mod keys;
 pub mod operator;
 pub mod params;
 pub mod prelude;
@@ -22,3 +21,6 @@ pub mod proof;
 pub mod relay;
 pub mod sorted_vec;
 pub mod utils;
+
+#[macro_use]
+mod macros;
